requirements.txt
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
pandas
paddlepaddle
numpy
pillow
opencv-python
matplotlib
flask
dual
PyYaml
seaborn
paddleocr==2.0.1

